23: Refactored by renamed epics funcs
* add auth integration service folder
* renamed epics and containers accordingly
