+++ front-end/services/dpdIntegration/auth/utils.js
@@ -1,0 +1,5 @@
+export function mapAuthUser(user) {
+    return Object.assign({}, user, {
+        email: user.username,
+    });
+}
