+++ front-end/services/dpdIntegration/index.js
@@ -1,9 +1,11 @@
+import * as auth from './auth/';
 import * as user from './user/';
 import * as page from './page/';
 import * as template from './template/';
 
 export default Object.assign(
     {},
+    auth,
     user,
     page,
     template,
