+++ front-end/services/dpdIntegration/user/index.js
@@ -1,56 +1,5 @@
-import co from 'co';
 import * as utils from './utils.js';
 import { dpdHandler, dpdDelHandler } from '../utils.js';
-
-export function login(email, password) {
-    return new Promise(function(resolve, reject) {
-        const promiseHandler = dpdHandler(resolve, reject);
-
-        dpd.users.login({
-            'username': email,
-            'password': password
-        }, promiseHandler(function dpdUsersLogin(me) {
-            return me;
-        }));
-    });
-}
-
-export function getAuthUser(email, password) {
-    return new Promise(function(resolve, reject) {
-        const promiseHandler = dpdHandler(resolve, reject);
-
-        dpd.users.me(
-            promiseHandler(function dpdUsersMe(me) {
-                return utils.mapAuthUser(me);
-            })
-        );
-    });
-}
-
-export function loginAndGetAuthUser(email, password) {
-    return co(function* () {
-        try {
-            yield login(email, password);
-            const user = yield getAuthUser();
-            return user;
-        } catch (e) {
-            throw e;
-        }
-    });
-}
-
-export function logout() {
-    return new Promise(function(resolve, reject) {
-        dpd.users.logout(function(err) {
-            if (err) {
-                reject(err);
-                return;
-            };
-
-            resolve();
-        });
-    });
-}
 
 export function getUsers() {
     return new Promise(function(resolve, reject) {
