+++ front-end/services/dpdIntegration/user/utils.js
@@ -1,9 +1,3 @@
-export function mapAuthUser(user) {
-    return Object.assign({}, user, {
-        email: user.username,
-    });
-}
-
 export function mapUser (user) {
     return {
         id: user.id,
